Restructure filesystem and modules, #44
