+++ maverick_api/data/parameters/ArduCopter.xml
