+++ maverick_api/data/static/graphiql.html
