+++ maverick_api/data/static/script.txt
