+++ maverick_api/modules/base/setup/config.py
@@ -1,13 +1,15 @@
 import json, os
 
-
 class Configuration(object):
-    def __init__(self, config_name):
-        self.read_config(config_name)
-
-    def read_config(self, config_name):
+    def __init__(self, config_name, config_file):
+        self.config_file = config_file
+        self.config_name = config_name
+        self.read_config()
+        
+            
+    def read_config(self):
         try:
-            with open(os.path.join(os.path.dirname(__file__), config_name), "r") as fid:
+            with open(self.config_file, 'r') as fid:
                 self.config = json.loads(fid.read())
         except Exception as e:
             # TODO: catch the correct error here...
@@ -15,6 +17,6 @@ class Configuration(object):
             self.config = None
             print("ERROR: {0}".format(e))
             raise SystemExit()
-
+    
     def get_config(self):
         return self.config
