+++ maverick_api/system/start_mongo.sh
